% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge-ID.R
\name{forge_ID}
\alias{forge_ID}
\title{forge_ID function}
\usage{
forge_ID(wd)
}
\arguments{
\item{wd}{Character: The working directory path where output files are stored.
The function will generate a unique ID based on the existing files in the "Outputs" subdirectory.}
}
\value{
A character string containing a unique ID to be used for saving files.
}
\description{
This function generates a unique ID by checking the output directory to ensure no duplicate IDs are created.
}
\examples{
# Example usage:
wd <- "~/Documents/LAB_ECO"
uniqueID <- forge_ID(wd)
}
