% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Saver.R
\name{Line_Saver}
\alias{Line_Saver}
\title{Line_Save Function}
\usage{
Line_Saver(params, uniqueID, wd)
}
\arguments{
\item{params}{List. A list containing elements generated by the \code{generate} function. This list should include the following elements:
\itemize{
\item Interactions: If \code{NULL}, a random interaction matrix is generated using \code{runif(n = n_species, min = 0, max = 1)}.
\item Growths: Numeric vector representing the growth rates of the simulated species.
\item Population: Numeric vector representing the initial abundances of the simulated species.
\item Seeds: Numeric vector representing the seeds used to generate the interaction matrix, growth rates, and populations.
}}

\item{uniqueID}{Character. A unique ID string used for saving files. This ID should be generated previously using the \code{generate_uniqueID} function.}

\item{wd}{Character. Path to the working directory where files will be saved. The file will be saved in the path: \code{wd/Parameters/P_uniqueID.tsv}.}
}
\value{
A list containing the generated ID and the path where the simulation file is stored.
}
\description{
This function saves parameters by lines.
}
\examples{
# Example usage

wd <- "~/Documents/LAB_ECO"

# Initial parameters
N_species <- 2
C0 <- 0.45
CN <- 0.2
Diag_val <- -0.5

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv")
params <- generate(N_species, seeds_path, C0, CN, Diag_val)

# Generate unique ID
uniqueID <- generate_uniqueID(wd)

# Save by lines
Line_Save(N_species, C0, CN, Diag_val, output_matrix, params)

}
