% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\name{output_save}
\alias{output_save}
\title{Output_save function}
\usage{
output_save(Output, uniqueID, wd)
}
\arguments{
\item{uniqueID}{A character string containing a unique ID to be used for saving files. Generated previously with \code{generate_uniqueID} function.}

\item{wd}{Character. Path to the working directory where files will be saved. The file will be saved in the path: \code{wd/Outputs/O_uniqueID.tsv}.}

\item{output}{A data frame or matrix containing the simulation output to be saved.}
}
\value{
A message indicating the ID and the file path is printed to the console.
}
\description{
This function saves the provided output data to a TSV file with a specified ID in a given working directory.
}
\details{
The function constructs the file path using the working directory (\code{wd}), the "Outputs" subdirectory,
and the unique identifier (\code{ID}). The file is saved with a ".tsv" extension where the columns represent generations
and rows species.
}
\examples{

wd <- "~/Documents/LAB_ECO"

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv" )
# It can also be: "~/Documents/LAB_ECO/Seeds.tsv"
params <- generate (N_species = 2, seeds_path, C0 = 0.45, CN = 0.2, Diag_val = -0.5)

# Simulate species interactions
times <- 100 # Define the number of generations
Output <- Simulate_output(N_species = 2, params = params, times = times, norm = FALSE)

# Generate unique ID
uniqueID <- generate_uniqueID(wd)

# Save output
output_save(Output, uniqueID, wd)
}
