% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS-find-and-save-all.R
\name{SS_find_and_save_all}
\alias{SS_find_and_save_all}
\title{SS_find_and_save_all Function}
\usage{
SS_find_and_save_all(uniqueID, output, tolerance, wd)
}
\arguments{
\item{uniqueID}{Character: A unique identifier for the simulation run.}

\item{output}{Matrix: Simulation results where rows represent species and columns represent time steps.}

\item{tolerance}{Numeric: Tolerance value for determining the steady state.}

\item{wd}{Character: Working directory where the results will be saved.}
}
\description{
This function searches for steady states in a simulation by analyzing the differences in the simulation results over time and saves the
generation where the steady state is reached.
}
\examples{
# Example usage:

wd <- "~/Documents/LAB_ECO" # Working directory

# Generate parameters
N_species = 2
C0 = 0.45
CN = 0.2
Diag_val = -0.5

seeds_path <- file.path(wd, "Seeds.tsv" )
params <-  init_data(N_species, seeds_path, C0, CN, Diag_val)

# Simulate species interactions
times <- 1000 # Define the number of generations
output <- Simulate_output(N_species = 2, params = params, times = times, norm = FALSE)

# Generate ID
uniqueID <- forge_ID(wd)


#--------------------------ALL  Steady states---------------------------#
tolerance = 0.05
SS_find_and_save_all(uniqueID, output, tolerance, wd)
}
