% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS-roll-window-all.R
\name{SS_roll_window_all}
\alias{SS_roll_window_all}
\title{SS_roll_window_all Function}
\usage{
SS_roll_window_all(output, tolerance)
}
\arguments{
\item{output}{Matrix: Simulation results where rows represent species and columns represent time steps.}

\item{tolerance}{Numeric: Tolerance value for determining the steady state.}
}
\value{
A list containing:
\itemize{
\item \code{Tolerance_trans}: Numeric. The tolerance value.
\item \code{Diff_mean}: Numeric vector. Column means of the squared differences between successive time steps, transformed using \code{log}.
\item \code{Stable_gen}: Numeric. The generation (time step) at which the system first reaches the steady state, based on the tolerance threshold.
}
}
\description{
This function searches for steady states in a simulation by analyzing the differences in the simulation results over time.
}
\examples{
# Example usage:

wd <- "~/Documents/LAB_ECO"

# Initial parameters
N_species = 2
C0 = 0.45
CN = 0.2
Diag_val = -0.5

# Generate parameters
seeds_path <- file.path(wd, "Seeds.tsv" )
params <-  init_data(N_species, seeds_path, C0, CN, Diag_val)

# Generate simulation
times <- 20 # Define the number of generations
output <- Simulate_output(N_species, params = params, times = times, norm = FALSE)

tolerance = 0.05
result <- SS_roll_window_all(output, tolerance)
}
